--- src/atomic_conf.h
@@ -1,1917 +1,0 @@
-/** \file atomic_conf.h
- *
- *  \brief List of neutral atomic configurations.
- */
-
-#ifndef __ATOMIC_CONF_H__
-#define __ATOMIC_CONF_H__
-
-const std::vector<std::vector<atomic_level_descriptor>> atomic_conf = {
-    { // H, z = 1
-        {1, 0, 1, 1.000000}
-    },
-    { // He, z = 2
-        {1, 0, 1, 2.000000}
-    },
-    { // Li, z = 3
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 1.000000}
-    },
-    { // Be, z = 4
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000}
-    },
-    { // B, z = 5
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 1.000000}
-    },
-    { // C, z = 6
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 1.000000},
-        {2, 1, 2, 1.000000}
-    },
-    { // N, z = 7
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 1.000000},
-        {2, 1, 2, 2.000000}
-    },
-    { // O, z = 8
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 2.000000}
-    },
-    { // F, z = 9
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 3.000000}
-    },
-    { // Ne, z = 10
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000}
-    },
-    { // Na, z = 11
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 1.000000}
-    },
-    { // Mg, z = 12
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000}
-    },
-    { // Al, z = 13
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 1.000000}
-    },
-    { // Si, z = 14
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 1.000000},
-        {3, 1, 2, 1.000000}
-    },
-    { // P, z = 15
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 1.000000},
-        {3, 1, 2, 2.000000}
-    },
-    { // S, z = 16
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 2.000000}
-    },
-    { // Cl, z = 17
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 3.000000}
-    },
-    { // Ar, z = 18
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000}
-    },
-    { // K, z = 19
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {4, 0, 1, 1.000000}
-    },
-    { // Ca, z = 20
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Sc, z = 21
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 1.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Ti, z = 22
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 2.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // V, z = 23
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 3.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Cr, z = 24
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 3.000000},
-        {3, 2, 3, 2.000000},
-        {4, 0, 1, 1.000000}
-    },
-    { // Mn, z = 25
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 3.000000},
-        {3, 2, 3, 2.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Fe, z = 26
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 2.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Co, z = 27
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 3.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Ni, z = 28
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 4.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Cu, z = 29
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 1.000000}
-    },
-    { // Zn, z = 30
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000}
-    },
-    { // Ga, z = 31
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 1.000000}
-    },
-    { // Ge, z = 32
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 1.000000},
-        {4, 1, 2, 1.000000}
-    },
-    { // As, z = 33
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 1.000000},
-        {4, 1, 2, 2.000000}
-    },
-    { // Se, z = 34
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 2.000000}
-    },
-    { // Br, z = 35
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 3.000000}
-    },
-    { // Kr, z = 36
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000}
-    },
-    { // Rb, z = 37
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Sr, z = 38
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {5, 0, 1, 2.000000}
-    },
-    { // Y, z = 39
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 1.000000},
-        {5, 0, 1, 2.000000}
-    },
-    { // Zr, z = 40
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 2.000000},
-        {5, 0, 1, 2.000000}
-    },
-    { // Nb, z = 41
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 3.000000},
-        {4, 2, 3, 1.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Mo, z = 42
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 3.000000},
-        {4, 2, 3, 2.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Tc, z = 43
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 1.000000},
-        {5, 0, 1, 2.000000}
-    },
-    { // Ru, z = 44
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 3.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Rh, z = 45
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 4.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Pd, z = 46
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000}
-    },
-    { // Ag, z = 47
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 1.000000}
-    },
-    { // Cd, z = 48
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000}
-    },
-    { // In, z = 49
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 1.000000}
-    },
-    { // Sn, z = 50
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 1.000000},
-        {5, 1, 2, 1.000000}
-    },
-    { // Sb, z = 51
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 1.000000},
-        {5, 1, 2, 2.000000}
-    },
-    { // Te, z = 52
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 2.000000}
-    },
-    { // I, z = 53
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 3.000000}
-    },
-    { // Xe, z = 54
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000}
-    },
-    { // Cs, z = 55
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 1.000000}
-    },
-    { // Ba, z = 56
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // La, z = 57
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 1.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Ce, z = 58
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 1.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 1.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Pr, z = 59
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Nd, z = 60
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {4, 3, 4, 1.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Pm, z = 61
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {4, 3, 4, 2.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Sm, z = 62
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {4, 3, 4, 3.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Eu, z = 63
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {4, 3, 4, 4.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Gd, z = 64
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 3.000000},
-        {4, 3, 4, 4.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 1.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Tb, z = 65
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 5.000000},
-        {4, 3, 4, 4.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Dy, z = 66
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 4.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Ho, z = 67
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 5.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Er, z = 68
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 6.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Tm, z = 69
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 7.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Yb, z = 70
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Lu, z = 71
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 1.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Hf, z = 72
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 2.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Ta, z = 73
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 3.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // W, z = 74
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 3.000000},
-        {5, 2, 3, 1.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Re, z = 75
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 3.000000},
-        {5, 2, 3, 2.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Os, z = 76
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 2.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Ir, z = 77
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 3.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Pt, z = 78
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 5.000000},
-        {6, 0, 1, 1.000000}
-    },
-    { // Au, z = 79
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 1.000000}
-    },
-    { // Hg, z = 80
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000}
-    },
-    { // Tl, z = 81
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 1.000000}
-    },
-    { // Pb, z = 82
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 1.000000},
-        {6, 1, 2, 1.000000}
-    },
-    { // Bi, z = 83
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 1.000000},
-        {6, 1, 2, 2.000000}
-    },
-    { // Po, z = 84
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 2.000000}
-    },
-    { // At, z = 85
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 3.000000}
-    },
-    { // Rn, z = 86
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000}
-    },
-    { // Fr, z = 87
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 1.000000}
-    },
-    { // Ra, z = 88
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Ac, z = 89
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Th, z = 90
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 2.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Pa, z = 91
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 2.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // U, z = 92
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 3.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Np, z = 93
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 3.000000},
-        {5, 3, 4, 1.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Pu, z = 94
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 3.000000},
-        {5, 3, 4, 3.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Am, z = 95
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 3.000000},
-        {5, 3, 4, 4.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Cm, z = 96
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 3.000000},
-        {5, 3, 4, 4.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Bk, z = 97
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 5.000000},
-        {5, 3, 4, 4.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Cf, z = 98
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 4.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Es, z = 99
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 5.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Fm, z = 100
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 6.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Md, z = 101
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 7.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // No, z = 102
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 8.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Lr, z = 103
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 8.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 1.000000},
-        {7, 0, 1, 2.000000}
-    },
-    { // Rf, z = 104
-        {1, 0, 1, 2.000000},
-        {2, 0, 1, 2.000000},
-        {2, 1, 1, 2.000000},
-        {2, 1, 2, 4.000000},
-        {3, 0, 1, 2.000000},
-        {3, 1, 1, 2.000000},
-        {3, 1, 2, 4.000000},
-        {3, 2, 2, 4.000000},
-        {3, 2, 3, 6.000000},
-        {4, 0, 1, 2.000000},
-        {4, 1, 1, 2.000000},
-        {4, 1, 2, 4.000000},
-        {4, 2, 2, 4.000000},
-        {4, 2, 3, 6.000000},
-        {4, 3, 3, 6.000000},
-        {4, 3, 4, 8.000000},
-        {5, 0, 1, 2.000000},
-        {5, 1, 1, 2.000000},
-        {5, 1, 2, 4.000000},
-        {5, 2, 2, 4.000000},
-        {5, 2, 3, 6.000000},
-        {5, 3, 3, 6.000000},
-        {5, 3, 4, 8.000000},
-        {6, 0, 1, 2.000000},
-        {6, 1, 1, 2.000000},
-        {6, 1, 2, 4.000000},
-        {6, 2, 2, 2.000000},
-        {7, 0, 1, 2.000000}
-    }
-};
-
-#endif
